backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
anthropic==0.18.1
email-validator==2.1.0
bcrypt==4.0.1
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
slowapi==0.1.9
